feat: add axis manager module
